pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "tests"]

[project]
name = "devops-ai-agent"
version = "0.1.0"
description = "A DevOps AI Agent built on AgentCore that uses MCP servers to manage AWS infrastructure, monitor resources, and report via Teams."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "DevOps Team" }]

dependencies = [
    "boto3>=1.35.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "moto[ec2,cloudwatch]>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "boto3-stubs[ec2,cloudwatch]>=1.35.0",
]
infra = [
    "aws-cdk-lib>=2.150.0",
    "constructs>=10.3.0",
]

[project.scripts]
mcp-aws-infra = "src.mcp_servers.aws_infra.server:main"
mcp-monitoring = "src.mcp_servers.monitoring.server:main"
mcp-teams = "src.mcp_servers.teams.server:main"

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests (deselect with '-m \"not unit\"')",
    "integration: Integration tests (deselect with '-m \"not integration\"')",
]
